requirements1.txt
meson==1.8.1
meson-python==0.18.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.7.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
executing==2.1.0
fastapi==0.112.0
ffmpy==0.4.0
filelock==3.13.1
Flask==3.0.0
Flask-MySQLdb==2.0.0
fonttools==4.53.1
fpdf==1.7.2
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gradio==4.40.0
gradio_client==1.2.0
greenlet==3.2.3
groq==0.26.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
joLib==0.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.3.25
langchain-core==0.3.64
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.45
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.15.1
nest-asyncio==1.6.0
numpy==2.2.6
orjson==3.10.6
packaging==24.1
pandas==2.3.0
parso==0.8.4
patsy==1.0.1
pillow==10.4.0
platformdirs==4.1.0
plotly==5.24.1
prompt_toolkit==3.0.47
protobuf==5.29.0
psutil==6.0.0
pure_eval==0.2.3
pyarrow==18.1.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9
pytz==2024.1
pywin32==310
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.21.0
ruff==0.5.6
scikit-learn==1.7.0